add Local Permission

+++ catalog/views.py
@@ -3,12 +3,12 @@ from flask import Flask, render_template, request, redirect, jsonify, url_for, f
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
+from database_setup import Base, Category, Item, User
 from flask import session as login_session
 
 from catalog import app
 
-engine = create_engine('sqlite:///catalog.db')
+engine = create_engine('sqlite:///catalogUser.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -25,12 +25,38 @@ def homepage():
 def categoryList(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category.id)
-    return render_template('category.html', category = category, category_id = category_id, items = items)
+    if 'username' not in login_session:
+        return render_template('category-public.html', category = category, category_id = category_id, items = items)
+    else:
+        return render_template('category.html', category = category, category_id = category_id, items = items)
+
+@app.route('/catalog/new', methods=['GET', 'POST'])
+def newCategory():
+    if 'username' not in login_session:
+        return redirect('login')
+    category = session.query(Category)
+    print request.method
+    if request.method == 'POST':
+        newCategory = Category(
+            name=request.form['name'],
+                           user_id=login_session['user_id'])
+        session.add(newCategory)
+        session.commit()
+        flash("new item created.")
+        return redirect(url_for('hompage'))
+    else:
+        return render_template('new_category.html')
+
 
 @app.route('/catalog/<int:category_id>/<int:item_id>')
 def itemDescription(category_id, item_id):
     item = session.query(Item).filter_by(id=item_id).one()
-    return render_template('item.html',  category_id = category_id, item_id = item_id, item = item)
+    creator = session.query(User).filter_by(id=item.user_id).one()
+    creator = creator.name
+    if 'username' not in login_session:
+        return render_template('item-public.html',  category_id = category_id, item_id = item_id, item = item, creator = creator)
+    else:
+        return render_template('item.html',  category_id = category_id, item_id = item_id, item = item)
 
 @app.route('/catalog/<int:category_id>/new', methods=['GET', 'POST'])
 def newItem(category_id):
@@ -42,7 +68,8 @@ def newItem(category_id):
     if request.method == 'POST':
         newItem = Item(
             name=request.form['name'], description=request.form[
-                           'description'], category_id=category_id)
+                           'description'], category_id=category_id,
+                           user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash("new item created.")
@@ -52,9 +79,11 @@ def newItem(category_id):
 
 @app.route('/catalog/<int:category_id>/<int:item_id>/edit', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
-    if 'username' not in login_session:
-        return redirect('login')
     editedItem = session.query(Item).filter_by(id=item_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if editedItem.user_id != login_session['user_id']:
+        return "<script> function myFunction() {alert('You are not authorized to edit this item!');} </script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -70,8 +99,10 @@ def editItem(category_id, item_id):
 @app.route('/catalog/<int:category_id>/<int:item_id>/delete', methods=['GET', 'POST'])
 def deleteItem(category_id, item_id):
     if 'username' not in login_session:
-        return redirect('login')
+        return redirect('/login')
     deleteItem = session.query(Item).filter_by(id=item_id).one()
+    if deleteItem.user_id != login_session['user_id']:
+        return "<script> function myFunction() {alert('You are not ]authorized to delete this item!');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(deleteItem)
         session.commit()
